clean up compile problems for debian, add static linking

+++ util.c
@@ -5,7 +5,7 @@
 void *calloc_or_die(char *name, size_t nitems, size_t size) {
   void *ret=calloc(nitems,size);
   if (ret==NULL) {
-    fprintf(stdout,"failed to allocate memory for %d items of %d bytes for %s!\n",nitems,size,name);
+    fprintf(stdout,"failed to allocate memory for %d items of %d bytes for %s!\n",(int)nitems,(int)size,name);
     exit(2);
   }
   return ret;
